requirements.txt
channels
daphne
numpy
pillow
websockets
